Ooookay, I'm pretty tired so this may be a bunch of rambling. I have put this thing through the steps and it is pycogent compatible. That includes all the weird little nuances like making sure words are plural for things that are lists, arrays, whatever basically containers. 80 character max per a line fix. Didn't have to worry about method or property names. I could go on. It works.
There is a ton of documentation that has been added. It still doesn't have script usage added, but I'm just so tired right now that I'm going to pass out if I continue.

Best and multicola still aren't in here. This isn't a big deal because best hasn't been unit tested and multicola... well it needs to be looked into more. 

THERE ARE STILL DISCREPANCIES WITH DFA AND PERMANOVA OUTPUT. THIS WILL BE LOOKED INTO.